(Refactor): Adds following directories: api, controls, views. This ensures loose coupling and separation of concerns
