+++ frontend/src/api/categories.js
@@ -1,0 +1,16 @@
+import React, { useEffect, useState } from "react";
+
+function Filter({ onCategorySelect }) {
+  const [categories, setCategories] = useState([]);
+  const [selected, setSelected] = useState("");
+
+  useEffect(() => {
+    fetch("https://dummyjson.com/products/categories")
+      .then((res) => res.json())
+      .then((data) => setCategories(data))
+      .catch((err) => console.error("Failed to fetch categories", err));
+  }, []);
+
+}
+
+export default Filter
