+++ frontend/src/views/ProductDetails.jsx
